+++ src/pyrite/game.py
@@ -17,7 +17,7 @@ from pyrite._camera.camera import BaseCamera
 from pyrite._rendering.ortho_projection import OrthoProjection
 from pyrite._rendering.viewport import Viewport
 from pyrite._services.camera_service import CameraServiceProvider as CameraService
-from pyrite._transform import transform_system
+from pyrite._systems import transform_system
 from pyrite.utils import threading
 
 if TYPE_CHECKING:
