+++ src/pyrite/transform.py
@@ -1,7 +1,7 @@
 import pyrite._transform.transform
 import pyrite._component.transform_component
-import pyrite._transform.transform_system
+import pyrite._systems.transform_system
 
 Transform = pyrite._transform.transform.Transform
 TransformComponent = pyrite._component.transform_component.TransformComponent
-TransformSystem = pyrite._transform.transform_system.TransformSystem
+TransformSystem = pyrite._systems.transform_system.TransformSystem
